feat: DEVOPS-1559 / DEVOPS-1552 z2 deployer upgrade should not upgrade the node in parallel (#1470)
