+++ z2/Cargo.toml
@@ -29,6 +29,7 @@ bitvec = "1.0.1"
 blsful = { git = "https://github.com/JamesHinshelwood/agora-blsful", branch = "update-blstrs"}
 clap = {version = "4.5.17", features = ["derive"]}
 clap-verbosity-flag = "2.2.1"
+cliclack = "0.3.5"
 colored = "2.1.0"
 crypto-bigint = "=0.5.5"
 env_logger = "0.11.5"
