+++ z2/src/lib.rs
@@ -9,7 +9,6 @@ pub mod docs;
 mod github;
 pub mod kpi;
 pub mod mitmweb;
-mod node;
 pub mod otel;
 pub mod otterscan;
 pub mod perf;
